OIDC keys: Rename to public_keys
